ref: move it to the CLI
