--- elasticms-admin/src/Commands.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Admin;
-
-class Commands
-{
-    public const string EMSADMIN_FAKE_PROJECT_BUILD = 'emsadmin:fake-project:build';
-}
