+++ elasticms-cli/src/Commands.php
@@ -22,4 +22,5 @@ class Commands
 
     final public const string MULTI_FILES_FIELD_TO_ARCHIVE = 'emscli:files-field:to-archive';
     final public const string DEAD_LINKS_REPORT = 'emscli:web:dead-links-report';
+    final public const string FAKE_PROJECT_BUILD = 'emscli:dev:fake-project-build';
 }
